Screen reader improvements
* Don't read caption
* Don't group texts because only part of them is read

There is still a problem with a link instead of a list item (e.g. after "list of tutorials" in the call for tutorials, the list item header is read again instead of the rest of the line). I used Mac VoiceOver. The Chrome screen reader is worse but I'm not sure it's due to problems in the website.